requirements.txt
deta==1.2.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.27.2
